fix: map Haystack kebab roles in CLI ingest so slice CI matches soak hours
fdd_cli ingest left fan-status/zone-air-temp unmapped; package ingest already
translated them. Share haystack_point_to_role and run slice rules with
confirm-seconds 0 like synthetic soak confirm_min=0.
